+++ dependency/core/applications/age_classification/__init__.py
@@ -1,3 +1,4 @@
 from .age_classification import AgeClassification as Classifier
+from .age_classification import AgeClassificationRoi as RoiClassifier
 
-__all__ = ["Classifier"]
+__all__ = ["Classifier", "RoiClassifier"]