+++ dependency/core/applications/category_identification/__init__.py
@@ -1,3 +1,4 @@
 from .category_identification import CategoryIdentification as Classifier
+from .category_identification import CategoryIdentificationRoi as RoiClassifier
 
-__all__ = ["Classifier"]
+__all__ = ["Classifier", "RoiClassifier"]