+++ dependency/core/applications/exposure_identification/__init__.py
@@ -1,3 +1,4 @@
 from .exposure_identification import ExposureIdentification as Classifier
+from .exposure_identification import ExposureIdentificationRoi as RoiClassifier
 
-__all__ = ["Classifier"]
+__all__ = ["Classifier", "RoiClassifier"]