+++ dependency/core/processor/__init__.py
@@ -1,5 +1,5 @@
 from .detector_processor import DetectorProcessor
 from .detector_tracker_processor import DetectorTrackerProcessor
 from .classifier_processor import ClassifierProcessor
+from .roi_classifier_processor import RoiClassifierProcessor
 from .processor_server import ProcessorServer
-